+++ source/global/helpers.R
@@ -93,7 +93,8 @@ milQuant_dowloadHandler <<- function(plot = "plot", ftype = "png") {
     filename = paste(format(Sys.Date(), "%Y%m%d"),
                      "_milQuant_plot.", ftype, sep = ""),
     content <- function(file) {
-      ggsave(file, plot = plot, device = ftype,
+      ggsave(file, plot = plot + Plot_Base_Theme + Plot_Base_Guide,
+             device = ftype,
              width = 25, height = 15, units = "cm")
     }
   )
@@ -116,9 +117,9 @@ make_layer_selector <<- function(data_all, inputId,
   # produce the selectInput object that will be displayed
   pickerInput(inputId = inputId,
               label = label,
-              choices = sort(selectable_layers),
+              choices = selectable_layers,
               multiple = TRUE,
-              selected = sort(selectable_layers),
+              selected = selectable_layers,
               options = list("actions-box" = TRUE,
                              "live-search" = TRUE,
                              "live-search-normalize" = TRUE,
@@ -155,3 +156,27 @@ mq_spinner <<- function(object) {
 
 
 
+convert_to_Plotly <<- function(ggplot_p,
+                               source = "ggplot_source",
+                               tooltip = c("y", "x", "fill")) {
+  # add theme
+  ggplot_p <- ggplot_p + Plot_Base_Theme
+
+  plot_ly <- ggplotly(ggplot_p, source = source, tooltip = tooltip) %>%
+    layout(title = list(text = paste0(ggplot_p$labels$title, "<br>",
+                                      "<sup>", ggplot_p$labels$subtitle, "</sup>")),
+           annotations = list(text = ggplot_p$labels$caption,
+                              xref = "paper", yref = "paper",
+                              xanchor="right", yanchor="top",
+                              x = 1, y = 1,
+                              showarrow = FALSE,
+                              bgcolor = "white",
+                              font = list(size = 10)),
+           yaxis = list(tickmode = "auto", showline = FALSE, gridwidth = 3,
+                        gridcolor = "grey20"),
+           xaxis = list(gridcolor = "grey60"),
+           showlegend = TRUE) %>%
+    config(displaylogo = FALSE,
+           modeBarButtonsToRemove = c("select2d", "lasso2d"))
+  return(plot_ly)
+}