+++ source/header_sidebar.R
@@ -5,7 +5,7 @@ header <- dashboardHeader(
   #                        height = 44)),
   title = "milQuant",
   tags$li(class = "dropdown",
-          actionButton(label = "Quit",
+          actionButton(label = "Save and Quit",
                        icon = icon("power-off"),
                        inputId = "close_app")))
 
@@ -18,7 +18,8 @@ sidebar <- dashboardSidebar(
                                         container = tags$p))),
     uiOutput("select_operation"),
     uiOutput("select_trench"),
-    menuItem("Overview (DB-Project)", tabName = "overview", icon = icon("graduation-cap")),
+    menuItem("Overview (Place)", tabName = "overview", icon = icon("graduation-cap")),
+    menuItem("Workflow", tabName = "workflow", icon = icon("gear")),
     menuItem("All Finds", tabName = "allfinds", icon = icon("chart-bar")),
     menuItem("Pottery", tabName = "pottery_all", icon = icon("trophy"),
           menuSubItem("Pottery (single)", tabName = "pottery",