+++ source/server/allfinds_serv.R
@@ -1,22 +1,27 @@
-findPlot_data <- reactive({
-  findPlot_data <- selected_db() %>%
+findPlot_base_data <- reactive({
+
+  validate(
+    need(is.data.frame(selected_db()), "No Trenches and/or Places selected.")
+  )
+
+  findPlot_base_data <- selected_db() %>%
     filter(type %in% find_types) %>%
     remove_na_cols() %>%
     inner_join(react_index()[,c("identifier", "Operation", "Place")],
                by = "identifier")
-  return(findPlot_data)
+  return(findPlot_base_data)
 })
 
 output$allfinds_n <- renderText({
   validate(
-    need(findPlot_data(), "No project selected.")
+    need(findPlot_base_data(), "No project selected.")
   )
-  prettyNum(nrow(findPlot_data()), big.mark = ",")
+  prettyNum(nrow(findPlot_base_data()), big.mark = ",")
 })
 
 output$allfinds_overview <- renderText({
-  n_objects <- nrow(findPlot_data())
-  n_layers <- length(unique(findPlot_data()$relation.liesWithinLayer))
+  n_objects <- nrow(findPlot_base_data())
+  n_layers <- length(unique(findPlot_base_data()$relation.liesWithinLayer))
   paste("The selected trenches ", paste(input$select_trench, collapse = ", "),
         " (from ", paste(input$select_operation, collapse = ", "),
         ") contain a total of ", n_objects,
@@ -29,17 +34,13 @@ output$findPlot_period_selector <- renderUI({
 })
 
 output$findPlot_layer_selector <- renderUI({
-  make_layer_selector(findPlot_data(),
+  make_layer_selector(findPlot_base_data(),
                       inputId = "findPlot_layer_selector")
 })
 
 output$findPlot_var_selector <- renderUI({
-  all_cols <- colnames(findPlot_data())
-  findPlot_vars <- c("type")
-  findPlot_vars <- c(findPlot_vars, "storagePlace")
-  findPlot_vars <- c(findPlot_vars, "date")
-  findPlot_vars <- c(findPlot_vars, "Operation")
-  findPlot_vars <- c(findPlot_vars, "Place")
+  all_cols <- colnames(findPlot_base_data())
+  findPlot_vars <- c("type", "storagePlace", "date", "Operation", "Place")
   findPlot_vars <- c(findPlot_vars, all_cols[grepl("relation", all_cols)])
   findPlot_vars <- c(findPlot_vars, all_cols[grepl("workflow", all_cols)])
   findPlot_vars <- c(findPlot_vars, all_cols[grepl("period", all_cols)])
@@ -61,33 +62,53 @@ output$findPlot_var_selector <- renderUI({
 
 })
 
-make_allFindsPlot <- reactive({
-  findPlot_tmp <- findPlot_data() %>%
+allFindsPlot_data <- reactive({
+  validate(
+    need(is.character(input$findPlot_PlotVar), "No variable selected.")
+  )
+  allFindsPlot_data <- findPlot_base_data() %>%
     filter(relation.liesWithinLayer %in% input$findPlot_layer_selector) %>%
     period_filter(is_milet = is_milet, selector = input$findPlot_period_selector)
 
+  return(allFindsPlot_data)
+})
+
+make_allFindsPlot <- reactive({
+  plot_data <- allFindsPlot_data() %>%
+    mutate(var = get(input$findPlot_PlotVar)) %>%
+    select(type, var) %>%
+    mutate(type = fct_infreq(type))
+
+  if (is.logical(plot_data$var)) {
+    plot_data <- plot_data %>%
+      mutate(var = ifelse(is.na(var),
+                          FALSE,
+                          TRUE))
+  } else {
+    plot_data <- plot_data %>%
+      mutate(var = fct_infreq(var))
+  }
+
+
   if (input$findPlot_axis == "var_is_fill") {
-    p <- findPlot_tmp %>%
-      ggplot(aes(x = reorder(type,
-                             type,
-                             function(x)-length(x)),
-                 fill = get(input$findPlot_PlotVar))) +
+    p <- plot_data %>%
+      ggplot(aes(x = type,
+                 fill = var)) +
       labs(fill = input$findPlot_PlotVar,
            x = "Type of Find")
+    p
     if (input$findPlot_PlotVar == "period") {
       p <- p + scale_fill_period()
     }
   } else if (input$findPlot_axis == "var_is_x") {
     if (input$findPlot_PlotVar == "date") {
-      p <- findPlot_tmp %>%
+      p <- plot_data %>%
         ggplot(aes(x = date,
                    fill = type)) +
         scale_x_date(name = "Date of Processing")
     } else {
-      p <- findPlot_tmp %>%
-        ggplot(aes(x = reorder(get(input$findPlot_PlotVar),
-                               get(input$findPlot_PlotVar),
-                               function(x)-length(x)),
+      p <- plot_data %>%
+        ggplot(aes(x = var,
                    fill = type)) +
         labs(fill = "Type of Find",
              x = input$findPlot_PlotVar)
@@ -96,19 +117,21 @@ make_allFindsPlot <- reactive({
 
 
   p <- p + geom_bar(position = input$findPlot_bars) +
-    Plot_Base_Theme + Plot_Base_Guide +
     labs(title = input$findPlot_title,
          subtitle = input$findPlot_subtitle,
          caption = paste("Total number of objects: ",
-                         nrow(findPlot_tmp), sep = ""))
+                         nrow(plot_data), sep = ""))
   p
 })
 
-output$allFindsPlot <- renderPlot({
-  make_allFindsPlot()
+output$allFindsPlot <- renderPlotly({
+  convert_to_Plotly(make_allFindsPlot())
 })
+
 
 output$allFindsPlot_png <- milQuant_dowloadHandler(plot = make_allFindsPlot(),
                                                 ftype = "png")
 output$allFindsPlot_pdf <- milQuant_dowloadHandler(plot = make_allFindsPlot(),
                                                 ftype = "pdf")
+
+