+++ source/server/bricks_serv.R
@@ -1,4 +1,8 @@
 bricks <- reactive({
+  validate(
+    need(is.data.frame(selected_db()), "No Trenches and/or Places selected.")
+  )
+
   bricks <- selected_db() %>%
     filter(type == "Brick") %>%
     remove_na_cols()
@@ -36,22 +40,24 @@ output$bricks_period_selector <- renderUI({
 make_bricksPlot_1 <- reactive({
   #fill_name <- names(fill_vars[which(fill_vars == input$lwPlot_1_fillvar)])
 
-  p <- bricks() %>%
+  plot_data <- bricks() %>%
     # filter by periods from the slider if config is milet
     period_filter(is_milet = is_milet, selector = input$bricks_period_selector) %>%
-    filter(relation.liesWithinLayer %in% input$bricks_layer_selector) %>%
+    filter(relation.liesWithinLayer %in% input$bricks_layer_selector)
+
+  p <- plot_data  %>%
     ggplot(aes(x = brickForm)) +#, fill = get(input$bricksPlot_1_fillvar))) +
     geom_bar() +#name = fill_name) +
     scale_y_continuous(name = "number of bricks") +
     scale_x_discrete(name = "type of brick") +
-    labs(title = input$bricksPlot_1_title, subtitle = input$bricksPlot_1_subtitle) +
-    Plot_Base_Theme +
-    Plot_Base_Guide
+    labs(title = input$bricksPlot_1_title,
+         subtitle = input$bricksPlot_1_subtitle,
+         caption = paste("Total Number of Objects:", nrow(plot_data)))
   p
 })
 
-output$bricksPlot_1 <- renderPlot({
-  make_bricksPlot_1()
+output$bricksPlot_1 <- renderPlotly({
+  convert_to_Plotly(make_bricksPlot_1())
 })