+++ source/server/potteryQB_serv.R
@@ -1,4 +1,8 @@
 potteryQB <- reactive({
+  validate(
+    need(is.data.frame(selected_db()), "No Trenches and/or Places selected.")
+  )
+
   potteryQB <- selected_db() %>%
     filter(type == "Pottery_Quantification_B") %>%
     remove_na_cols()
@@ -24,7 +28,10 @@ output$QBpotPlot_1_period_selector <- renderUI({
   make_period_selector(inputId = "QBpotPlot_1_period_selector")
 })
 
-QBpotPlot_1 <- function() {
+QBpotPlot_1_data <- reactive({
+  validate(
+    need(is.data.frame(potteryQB()), "Data not available.")
+  )
   existing_cols <- colnames(potteryQB())
   keep <- existing_cols
   keep <- keep[grepl("count", keep)]
@@ -40,17 +47,25 @@ QBpotPlot_1 <- function() {
 
   plot_data <- plot_data %>%
     filter(relation.liesWithinLayer %in% input$QB_layer_selector) %>%
+    period_filter(is_milet = is_milet,
+                  selector = input$QBpotPlot_1_period_selector) %>%
     melt(id = c("identifier", "relation.liesWithinLayer", "period", "period.start", "period.end")) %>%
     mutate(value = ifelse(is.na(value), 0, value)) %>%
     mutate(value = as.numeric(value)) %>%
     mutate(variable = gsub("count", "", variable)) %>%
     mutate(variable = gsub("Rim|Base|Handle|Wall", "", variable)) %>%
     uncount(value) %>%
-    period_filter(is_milet = is_milet, selector = input$QBpotPlot_1_period_selector)
+    mutate(variable = fct_infreq(variable))
+  return(plot_data)
+})
+
+QBpotPlot_1 <- reactive({
+  validate(
+    need(is.data.frame(QBpotPlot_1_data()), "Data not available.")
+  )
 
   if (input$QBpotPlot_1_title == "") {
-    plot_title <- paste("Vessel Forms from ", input$select_operation,
-                        " in Context: ",
+    plot_title <- paste("Vessel Forms from Context: ",
                         paste(input$QB_layer_selector, collapse = ", "),
                         sep = "")
   } else {
@@ -59,17 +74,13 @@ QBpotPlot_1 <- function() {
 
 
   if (input$QBpotPlot_2_display == "function") {
-    p <- ggplot(plot_data, aes(x = fct_infreq(variable),
+    p <- ggplot(QBpotPlot_1_data(), aes(x = variable,
                                fill = period)) +
-      geom_bar(position = input$QBpotPlot_1_bars) +
-      Plot_Base_Theme +
       labs(x = "Vessel Forms", y = "count") +
       scale_fill_period(ncol = 9)
   } else if (input$QBpotPlot_2_display == "period") {
-    p <- ggplot(plot_data, aes(x = period,
-                               fill = fct_infreq(variable))) +
-      geom_bar(position = input$QBpotPlot_1_bars) +
-      Plot_Base_Theme + Plot_Base_Guide +
+    p <- ggplot(QBpotPlot_1_data(), aes(x = period,
+                               fill = variable)) +
       scale_fill_discrete(name = "Function", guide = "legend") +
       labs(x = "Period", y = "count")
   }
@@ -81,12 +92,14 @@ QBpotPlot_1 <- function() {
 
   p <- p +
     labs(title = plot_title,
-         subtitle = input$QBpotPlot_1_subtitle)
+         subtitle = input$QBpotPlot_1_subtitle,
+         caption = paste("Total Number of Fragments:", nrow(QBpotPlot_1_data()))) +
+    geom_bar(position = input$QBpotPlot_1_bars)
   p
-}
+})
 
-output$QBpotPlot_1 <- renderPlot({
-  QBpotPlot_1()
+output$QBpotPlot_1 <- renderPlotly({
+  convert_to_Plotly(QBpotPlot_1())
 })