+++ source/tabs/overview_tab.R
@@ -6,10 +6,10 @@ overview_tab <- tabItem(
     infoBox(
       title = "Info",
       value = "This plot gives you an overview of all the resources that
-        exist in the Operation and/or Trenches you have selected in the sidebar
+        exist in the Place you have selected in the sidebar
         on the left. All other tabs accesible via the bar on the left will only
         display resources from the Operation and/or Trenches you
-        have selected.",
+        have selected in the second dropdown below.",
       icon = icon("vial"),
       color = "light-blue",
       width = 10),
@@ -22,6 +22,6 @@ overview_tab <- tabItem(
   fluidRow(
     box(title = "Overview", status = "primary",
         solidHeader = TRUE, collapsible = FALSE,
-        width = 12, height = 600,
-        plotOutput("overview") %>% mq_spinner()))
+        width = 12, height = 650,
+        plotlyOutput("overview", height = 570) %>% mq_spinner()))
 )