+++ source/tabs/workflow_tab.R
@@ -1,0 +1,8 @@
+workflow_tab <- tabItem(
+  tabName = "workflow",
+  title = "Workflow",
+
+  fluidRow(
+    uiOutput("workflow_tabs") %>% mq_spinner()
+    )
+)